% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LogisticSPSSlike.R
\name{LogisticSPSSlike}
\alias{LogisticSPSSlike}
\title{To keep the variable names in accordance with SPSS output when performing logistic regression using R}
\usage{
LogisticSPSSlike(fit)
}
\arguments{
\item{fit}{an object of class inheriting from "glm".}
}
\value{
a data frame contains B, Wald, OR, CI, and P values.
}
\description{
To keep the variable names in accordance with SPSS output when performing logistic regression using R
}
\examples{
if (!require("AER")) install.packages("AER")
library("AER")
data(Affairs)
Affairs[Affairs$affairs>1,]$affairs<-1
fit<-glm(affairs~.,data=Affairs,family=binomial())
formatFit(fit)

}
